Split out geofency with a component and platform
